Documented delete methods.
Documented the delete methods I did and deleted the methods that were used to store deleted data and register it as they were just used for testing.

the delete calls in search are currently commented so your databases don't get ruined by accident, uncomment it to test delete if you want to.